Separate out GUI specific utils from general ones
