+++ damnit/util.py
@@ -1,20 +1,12 @@
 import glob
 import sys
 from datetime import datetime, timezone
-from enum import Enum
-from pathlib import Path
 
 import numpy as np
 import pandas as pd
 from pandas.api.types import infer_dtype
-from sklearn.neighbors import KernelDensity
 
 from .context import add_to_h5_file
-
-
-class StatusbarStylesheet(Enum):
-    NORMAL = "QStatusBar {}"
-    ERROR = "QStatusBar {background: red; color: white; font-weight: bold;}"
 
 
 def timestamp2str(timestamp):
@@ -24,13 +16,6 @@ def timestamp2str(timestamp):
     dt_utc = datetime.fromtimestamp(timestamp, timezone.utc)
     dt_local = dt_utc.astimezone()
     return dt_local.strftime("%H:%M:%S %d/%m/%Y")
-
-
-def icon_path(name):
-    """
-    Helper function to get the path to an icon file stored under ico/.
-    """
-    return str(Path(__file__).parent / "gui/ico" / name)
 
 
 def make_finite(data):
@@ -59,23 +44,6 @@ def delete_variable(db, name):
             if name in f:
                 del f[f".reduced/{name}"]
                 del f[name]
-
-
-def kde(x: np.ndarray, npoints: int = 1000) -> tuple[np.ndarray, np.ndarray]:
-    """1D kernel density estimation.
-
-    Args:
-        x (np.ndarray): 1D array of data points.
-        npoints (int, optional): Number of points to evaluate the KDE at. Defaults to 1000.
-
-    Returns:
-        tuple[np.ndarray, np.ndarray]: Tuple of (x, y) coordinates of the KDE.
-    """
-    xplot = np.linspace(x.min(), x.max(), npoints)[:, np.newaxis]
-    kde = KernelDensity().fit(x[:, None])
-    log_dens = kde.score_samples(xplot)
-    y = np.exp(log_dens)
-    return xplot.squeeze(), y
 
 
 def isinstance_no_import(obj, mod: str, cls: str):
